fix: current buffer detection
